KAFKA-10770: Remove duplicate defination of Metrics#getTags
